(Example) Refactor font handling details into the new FontService class.
